Update scripts and add github-copilot-cli (#92)

+++ gentoo/setup.sh
@@ -1,27 +1,29 @@
 #!/usr/bin/env zsh
 
+set -e
+cd ~
+
 # Gentoo
+
 sudo emerge app-editors/vim app-misc/jq app-shells/zsh app-shells/zsh-syntax-highlighting gentoo-zsh-completions dev-vcs/git dev-lang/python \
 dev-python/pip dev-lang/ruby dev-lang/go app-admin/helm app-admin/terraform sys-cluster/kubectl dev-vcs/pre-commit \
-app-admin/vault app-admin/kubectx dev-util/github-cli
+app-admin/vault app-admin/kubectx dev-util/github-cli net-libs/nodejs
 
-# Go
+# Terraform Docs
+
 go install github.com/terraform-docs/terraform-docs@v0.16.0
 
 # Infracost
+
 curl -fsSL https://raw.githubusercontent.com/infracost/infracost/master/scripts/install.sh | sh
 
 # Automatically enable pre-commit on repositories
+
 git config --global init.templateDir ~/.git-template
 pre-commit init-templatedir ~/.git-template
 
-# Pull with Rebase
-git config --global pull.rebase true
-
-# Prune on Fetch
-git config --global fetch.prune true
-
 # Ruby Tools
+
 export RUBYOPT="-W:no-deprecated -W:no-experimental"
 
 echo 'gem: --no-document' > ~/.gemrc
@@ -40,12 +42,13 @@ Style/FrozenStringLiteralComment:
 EOF
 
 # Pathogen.vim
+
 mkdir -p ~/.vim/autoload ~/.vim/bundle
 curl -LSso ~/.vim/autoload/pathogen.vim https://tpo.pe/pathogen.vim
 
 # Pathogen Plugins
-cd ~/.vim/bundle
-git clone https://github.com/hashivim/vim-terraform.git
+
+git clone https://github.com/hashivim/vim-terraform ~/.vim/bundle/vim-terraform
 
 cat << EOF > ~/.vimrc
 set visualbell
@@ -57,20 +60,31 @@ let g:terraform_align=1
 EOF
 
 # Google Cloud SDK
+
 curl https://sdk.cloud.google.com > install.sh
 bash install.sh --disable-prompts
 
 # Oh My Zsh
+
 sh -c "$(curl -fsSL https://raw.github.com/ohmyzsh/ohmyzsh/master/tools/install.sh)" "" --unattended
 
 # zsh-autosuggestions
+
 git clone https://github.com/zsh-users/zsh-autosuggestions ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-autosuggestions
 
+# GitHub Copilot CLI
+
+npm install @githubnext/github-copilot-cli
+
 # Powerlevel10k
+
 git clone --depth=1 https://github.com/romkatv/powerlevel10k.git ${ZSH_CUSTOM:-$HOME/.oh-my-zsh/custom}/themes/powerlevel10k
 
 # Shell Setup
-cp ~/.zshrc ~/.zshrc-`date +"%Y%m%d_%H%M%S"`.bak
+
+if [ -f ~/.zshrc ]; then
+    cp ~/.zshrc ~/.zshrc-`date +"%Y%m%d_%H%M%S"`.bak
+fi
 
 cat << 'EOF' >> ~/.zshrc
 alias gpg-passphrase="echo "test" | gpg --clearsign > /dev/null 2>&1"
@@ -97,11 +111,13 @@ source ~/google-cloud-sdk/path.zsh.inc
 
 autoload -U +X bashcompinit && bashcompinit
 complete -o nospace -C /usr/bin/terraform terraform
+eval "$(github-copilot-cli alias -- "$0")"
 EOF
 
-echo -e "export PATH=\$HOME/bin:\$HOME/.go/bin:\$HOME/.gem/bin:\$HOME/.local/bin:\$PATH\n$(cat ~/.zshrc)" > ~/.zshrc
+echo -e "export PATH=\$HOME/bin:\$HOME/.go/bin:\$HOME/.gem/bin:\$HOME/.local/bin:\$HOME/node_modules/.bin:\$PATH\n$(cat ~/.zshrc)" > ~/.zshrc
 
 # Create Update Script
+
 mkdir -p ~/bin
 cat << 'EOF' > ~/bin/update.sh
 #!/usr/bin/env zsh
@@ -133,6 +149,9 @@ git pull
 
 # Infracost
 curl -fsSL https://raw.githubusercontent.com/infracost/infracost/master/scripts/install.sh | sh
+
+# NPM
+npm update
 EOF
 
 chmod 755 ~/bin/update.sh
